requirements.txt
annotated-types==0.7.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
circuitbreaker==2.1.3
cryptography==44.0.3
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
oci==2.158.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
rich==14.1.0
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
